requirements.txt
exifread>=3.0
matplotlib
numpy
rawpy
OpenImageIO
pyimgui
pyglet
